rework browsing tool to use new submodule framework for readability
